Update to Gitea 1.24.0
Gitea 1.24.0 is coincidently built with Go v1.24.
I thus had to upgrade both the temporary building
and runtime source images. The build image was
easy, just update a version number. The runtime
image was not so easy.

Balena no longer updates the balenalib/armv7hf-alpine
image, with the logic that the normal alpine image
is now good enough. So I have converted to using
the base image, which required some tweaks to
configuration.