fix(#4): Build for correct arm architectures
As the name implies, balenalib/armv7hf-alpine was only built for the
ARM v7 platform. Unfortunately, when I switched to the alpine image,
they publish images for many architectures. Since I build on a ARM 64
device, that means that the 1.24.0 and 1.24.2 images were mistakenly
updated to ARM 64, making them incompatible with older 32 bit devices
and OSes (see issue #4).

This commit parameterizes the build process to support building a
variety of architectures, starting with arm v6, arm v7, and arm 64.